backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
cachecontrol==0.14.1
    # via firebase-admin
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
cryptography==43.0.3
    # via pyjwt
firebase-admin==6.6.0
    # via -r requirements.in
google-api-core[grpc]==2.22.0
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.151.0
    # via firebase-admin
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-core==2.4.1
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.19.0
    # via
    #   -r requirements.in
    #   firebase-admin
google-cloud-storage==2.18.2
    # via firebase-admin
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.67.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.67.1
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
msgpack==1.1.0
    # via cachecontrol
proto-plus==1.25.0
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==5.28.3
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via -r requirements.in
pydantic-core==2.23.4
    # via pydantic
pyjwt[crypto]==2.9.0
    # via firebase-admin
pyparsing==3.2.0
    # via httplib2
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
